graph_processing/pyproject.toml
[tool.poetry]
name = "graph-processing"
version = "0.1.0"
description = ""
authors = ["Mariusz Paluch <mariuszpaluch001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
faust-streaming = "^0.11.3"
networkx = "^3.3"
simple-settings = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
